:sparkles: Add sync ranks command

+++ commands/admin/sync-ranks.js
@@ -27,11 +27,17 @@ module.exports = class extends Command {
             return message.success("admin/sync-ranks:BEING_CANCELLED");
         }
 
+        const [guildRanks, membersData] = await Promise.all([
+            this.client.database.fetchGuildRanks(message.guild.id),
+            this.client.database.fetchGuildLeaderboard(message.guild.id, message.guild.settings.storageID)
+        ]);
+
         const conf = await message.sendT("admin/sync-ranks:CONFIRM", {
             success: this.client.config.emojis.success,
             error: this.client.config.emojis.error
         });
         await message.channel.awaitMessages((m) => m.author.id === message.author.id && (m.content === "cancel" || m.content === "-confirm"), { max: 1, time: 90000 }).then(async (collected) => {
+            if (!collected.first()) return;
             if (collected.first().content === "cancel") return conf.error("common:CANCELLED", null, true);
             collected.first().delete().catch(() => {});
             conf.success("admin/sync-ranks:STARTED", {
@@ -40,32 +46,18 @@ module.exports = class extends Command {
                 edit: true
             });
 
-            await message.guild.members.fetch();
-            const members = message.guild.members.cache.array();
-
             this.client.syncRanksTasks[message.guild.id] = {
                 now: 10,
-                total: members.length,
+                total: membersData.length,
                 status: "RUNNING"
             };
 
-            console.time("fetch members");
-            await this.client.database.fetchMembers(message.guild.members.cache.map((m) => {
-                return {
-                    userID: m.id,
-                    guildID: message.guild.id
-                };
-            }));
-            console.timeEnd("fetch members");
-
-            console.time("add roles");
-            while (members.length > 0 && (this.client.syncRanksTasks[message.guild.id] && this.client.syncRanksTasks[message.guild.id].status === "RUNNING")){
-                const member = members.shift();
+            while (membersData.length > 0 && (this.client.syncRanksTasks[message.guild.id] && this.client.syncRanksTasks[message.guild.id].status === "RUNNING")){
+                const memberData = membersData.shift();
                 this.client.syncRanksTasks[message.guild.id].now++;
-                const memberData = await this.client.database.fetchMember(member.id, member.guild.id);
-                await this.client.functions.assignRanks(member, memberData.calculatedInvites, data.guild.ranks, data.guild.keepRanks, data.guild.stackedRanks).catch(() => {});
+                const member = message.guild.members.cache.get(memberData.userID) || await message.guild.members.fetch(memberData.userID).catch(() => {});
+                await this.client.functions.assignRanks(member, memberData.invites, guildRanks, message.guild.settings.keepRanks, message.guild.settings.stackedRanks).catch(() => {});
             }
-            console.timeEnd("add roles");
 
             if (this.client.syncRanksTasks[message.guild.id] && this.client.syncRanksTasks[message.guild.id].status === "RUNNING"){
                 message.success("admin/sync-ranks:SUCCESS");
